add folder for localization and divide types into logic parts
