+++ CookBook/app/_layout.tsx
@@ -7,7 +7,7 @@ import { useEffect } from 'react'
 import 'react-native-reanimated'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 const queryClient = new QueryClient()
-import '@/services/i18n'
+import '@/localization/i18n'
 import * as Sentry from '@sentry/react-native'
 import '@/services/sentry'
