+++ CookBook/components/ButtonLogin.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { TouchableOpacity, Text, StyleSheet } from 'react-native'
-import { ButtonRegisterProps } from '@/types/types'
+import { ButtonRegisterProps } from '@/types/button'
 import { useTranslation } from 'react-i18next'
 import { COLORS, LAYOUT } from '@/constants/Constants'
