+++ CookBook/components/__tests__/ButtonQuit.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { fireEvent, render } from '@testing-library/react-native'
 import { ButtonQuit } from '@/components/ButtonQuit'
-import i18n from '@/services/i18n'
+import i18n from '@/localization/i18n'
 describe('', () => {
   test('button renders with correct text', () => {
     const { getByText } = render(<ButtonQuit onPress={() => {}} />)
