+++ CookBook/localization/i18n.ts
@@ -1,7 +1,7 @@
 import i18n from 'i18next'
 import { initReactI18next } from 'react-i18next'
-import enTranslation from '@/locales/en.json'
-import ruTranslation from '@/locales/ru.json'
+import enTranslation from '@/localization/locales/en.json'
+import ruTranslation from '@/localization/locales/ru.json'
 i18n.use(initReactI18next).init({
   resources: {
     en: { translation: enTranslation },
