+++ CookBook/package.json
@@ -24,7 +24,6 @@
     "eas-cli": "^16.18.0",
     "expo": "~53.0.22",
     "expo-constants": "^17.1.7",
-    "expo-dev-client": "~5.2.4",
     "expo-font": "~13.3.2",
     "expo-linking": "~7.1.7",
     "expo-router": "~5.1.5",
