+++ CookBook/types/auth.ts
@@ -1,0 +1,9 @@
+export type PropsEmail = {
+  value: string
+  onChangeText: (text: string) => void
+}
+
+export type PropsPassword = {
+  value: string
+  onChangeText: (text: string) => void
+}
