+++ CookBook/types/button.ts
@@ -1,0 +1,9 @@
+export type ButtonRegisterProps = {
+  onPress: () => void
+  isLoading: boolean
+  disabled: boolean
+}
+
+export type ButtonQuitProps = {
+  onPress: () => void
+}
